requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
requests==2.31.0
pytz==2023.3
apscheduler==3.10.4
